backend/requirements.txt
Flask
google-genai
python-dotenv
flask-cors
